Fixed exiting the game not transitioning into Main Menu, re-implemented crossfading into levels, GameManager gets current level on Start (helpful for testing)
It looks like crossfading between scenes got messed up after the UI got reworked, so I readded the Crossfade GameObject to the Game_UI prefab so we can crossfade between scenes.
Also made the GameManager get the currentLevel in its Start() function. This won't bring in anything noticable, but previously if we were to load a given level that we can play on, beat it, and try to go to the next, it would bring us back to the main menu. Now, it properly goes into the next level.