Fixed error in DestroyWithTag script, updated UI crossfade to account for Tetris spot
DestroyOnTag - Update() function's if statement didn't check if we had a can inside the trigger before trying to access the Item script on it. This was causing a non-fatal error, but an error nonetheless.

UI - "Reserved" text was not on being rendered below the crossfade and would stand out upon loading into and loading out of a scene.

Animation - The start button animation is there (it just flashes), but I have it disabled b/c I don't actually like how it looks. To enable it, click on the Start Button and enable its Animator.