Using the tetris spot system wasn't working, polished tutorial and fixed bugs with it, fixed level selection bug where spamming can load the level many times.
Tetris holding spot wasn't working - items would disappear upon being pickup up. I polished the code so now we only have to deal with the Pickup script and the itemHeld variable.

Tutorial wasn't updating text properly, so I polished it up while using some of the old code to make it easier to read and "safer".

Made it so crossfade image (the big black box that covers the screen) properly blocks raycasts (mouse clicks in this case). This prevents spamming and loading a lvl mulitple times.